Refactored error tests and begin work on a real IT test
